Removed folders and improved aug
